Day shift

+++ server/app/main.py
@@ -8,6 +8,7 @@ from werkzeug.utils import secure_filename
 
 from entities import tasks, challenges
 from entities import mongo_connect as db
+from entities.tasks import Path, Task
 import utils
 
 app = Flask(__name__)
@@ -39,29 +40,32 @@ def task_find(obj_id):
         return abort(404)
 
 
-@app.route("/task/create")
+@app.route("/task/create", methods=["POST"])
 def add_task():
-    path = request.args.get('path')
-    name = request.args.get('name')
-    meta = request.args.get('meta')
+    question, challenge, path, meta = utils.extract_args(request.args, ('question', 'challenge', 'path'))
 
+    if challenge:
+        path = challenges.Challenge.get_by_id(challenge, db_conn).path
+    if not path:
+        abort(400, 'Neither challenge, nor path provided')
     path = tasks.Path.get_by_id(path, db_conn)
-    task = tasks.Task(name=name, meta=meta)
+    task = tasks.Task(question=question, meta=meta)
     path.add_task(task, db_conn)
-    return 'ok'
+    return str(task._id)
 
 
 @app.route("/paths")
 def paths_all():
     return jsonify([p.view() for p in tasks.Path.get_all(db_conn)])
 
-@app.route("/path/create")
+@app.route("/path/create", methods=["POST"])
 def add_path():
     name = request.args.get('name')
     meta = request.args.get('meta')
 
-    tasks.Path(name=name, meta=meta).upload(db_conn)
-    return 'ok'
+    path = tasks.Path(name=name, meta=meta)
+    path.upload(db_conn)
+    return str(path._id)
 
 @app.route("/path/<obj_id>")
 def path_find(obj_id):
@@ -73,32 +77,22 @@ def path_find(obj_id):
 
 @app.route("/path/<obj_id>", methods=['DELETE'])
 def path_del(obj_id):
-    # challenges.Challenge.get_by_criteria({'path': ObjectId(obj_id)})
+    cascade = challenges.Challenge.get_by_criteria({'path': ObjectId(obj_id)}, db_conn)
+    if cascade:
+        abort(400, f'There still exist challenges: {[str(c._id) for c in cascade]}')
     res = db_conn.paths.delete_one({'_id': ObjectId(obj_id)})
     if res.deleted_count:
         return 'deleted'
     else:
         return abort(404)
 
-@app.route("/challenge/create")
+@app.route("/challenge/create", methods=["POST"])
 def add_challenge():
-    path = request.args.get('path')
-    student = request.args.get('student')
-    teacher = request.args.get('teacher')
-    meta = request.args.get('meta')
+    path, passphrase, meta = utils.extract_args(request.args, ('path', 'passphrase'))
 
-    ch = challenges.Challenge(path=ObjectId(path), meta=meta, student=student, teacher=teacher)
+    ch = challenges.Challenge(path=ObjectId(path), meta=meta, passphrase=passphrase)
     ch.upload(db_conn)
-    return 'ok'
-
-@app.route("/challenges")
-def challenge_list():
-    student_id = request.args.get('for')
-    if not student_id:
-        return abort(400, 'Need to specify "for" parameter')
-
-    res = challenges.Challenge.get_by_criteria({'student': student_id}, db_conn)
-    return jsonify([ch.view() for ch in res])
+    return str(ch._id)
 
 @app.route("/challenge/<obj_id>")
 def challenge_find(obj_id):
@@ -108,7 +102,7 @@ def challenge_find(obj_id):
     else:
         return abort(404, 'Challenge not found')
 
-@app.route("/challenge/<obj_id>", methods=['DELETE'])
+@app.route("/quest/<obj_id>", methods=['DELETE'])
 def challenge_del(obj_id):
     res = db_conn.challenges.delete_one({'_id': ObjectId(obj_id)})
     if res.deleted_count:
@@ -116,6 +110,57 @@ def challenge_del(obj_id):
     else:
         return abort(404)
 
+@app.route("/quest/<obj_id>")
+def quest_find(obj_id):
+    res = challenges.Challenge.get_by_id(obj_id, db_conn, with_tasks=True)
+    if res:
+        return jsonify(res)
+    else:
+        return abort(404, 'Challenge not found')
+
+@app.route("/quest/create", methods=["POST"])
+def add_quest():
+    path_name, passphrase, meta = utils.extract_args(request.args, ('name', 'passphrase'))
+    path = Path(name=path_name)
+    path.upload(db_conn)
+
+    meta['name'] = path_name
+    ch = challenges.Challenge(path=path._id, meta=meta, passphrase=passphrase)
+    ch.upload(db_conn)
+    return str(ch._id)
+
+@app.route("/quests")
+def challenge_list():
+    filters = dict(request.args)
+    res = challenges.Challenge.get_by_criteria(filters, db_conn)
+    return jsonify([ch.view() for ch in res])
+
+
+@app.route("/challenge/<obj_id>/check", methods=["POST"])
+def check_answer(obj_id):
+    required_params = ('task_id', 'answer')
+    task_id, answer, meta = utils.extract_args(request.args, required_params)
+
+    if any(p is None for p in (task_id, answer)):
+        abort(400, f"Missing one of the required parameters: {required_params}")
+
+    ch = challenges.Challenge.get_by_id(obj_id, db_conn, with_tasks=True)
+    if not ch:
+        abort(404, 'Challenge not found')
+
+    task = next((t for t in ch['tasks'] if t['id'] == task_id), None)
+    if task is None:
+        abort(404, 'Task not found in this challenge')
+
+    correct_answer = task.get('answer') or task.get('correct_option')
+    success = correct_answer == str(answer)
+    if success:
+        ch = challenges.Challenge.get_by_id(obj_id, db_conn)
+        if task['id'] not in ch.done:
+            ch.done.append(task['id'])
+            ch.update(db_conn, 'done')
+
+    return jsonify(success)
 
 @app.route("/upload", methods=['POST'])
 def handle_upload():
@@ -154,3 +199,48 @@ def route_frontend(path):
 if __name__ == "__main__":
     # Only for debugging while developing
     app.run(host="0.0.0.0", debug=True, port=8080, threaded=True)
+
+#
+# from flask import current_app, request, render_template, redirect, url_for
+# from flask.ext.login import LoginManager, login_user, logout_user, login_required, current_user
+# from wtforms import Form, StringField, PasswordField, validators
+#
+#
+# class User(flask_login.UserMixin):
+#     pass
+#
+# @login_manager.user_loader
+# def load_user(userid):
+#     return User.query.get(userid)
+#
+#
+# class UserLoginForm(Form):
+#     username = StringField('Username', [validators.DataRequired(), validators.Length(min=4, max=25)])
+#     password = PasswordField('Password', [validators.DataRequired(), validators.Length(min=6, max=200)])
+#
+#
+# @current_app.route('/login', methods=['GET', 'POST'])
+# def login():
+#     form = UserLoginForm(request.form)
+#     error = None
+#     if request.method == 'POST' and form.validate():
+#         user = User.query.filter_by(username=username.lower()).first()
+#         if user:
+#             if login_user(user):
+#                 current_app.logger.debug('Logged in user %s', user.username)
+#                 return redirect(url_for('secret'))
+#         error = 'Invalid username or password.'
+#     return render_template('login.html', form=form, error=error)
+#
+#
+# @current_app.route('/logout', methods=['GET'])
+# @login_required
+# def logout():
+#     logout_user()
+#     return redirect(url_for('login'))
+#
+#
+# @current_app.route('/secret', methods=['GET'])
+# @login_required
+# def secret():
+#     return 'This is a secret page. You are logged in as {}'.format(current_user.username)
